Modificação dos controlles para a nova versão
