+++ App/Views/admin/admins.phtml
@@ -1,6 +1,9 @@
 <div class="content p-4" id="musicas">  
         <div class="row">
             <div class="col-md-12">
+                <div class="form-group col-md-12 d-flex justify-content-end mt-2 mb-5">
+                    <a href="/cadastro-admin" class="btn btn-dark ml-2">Cadastrar Administrador</a>
+                </div>  
                 <div class="card mb-4 shadow ronded">
                     <div class="card-body">
                     <div class="col-md-12">
