+++ App/Views/admin/cadastro.phtml
@@ -4,7 +4,7 @@
             <h2>Cadastrar administrador</h2>
             <div class="card mb-4 mt-4 shadow ronded">
                 <div class="card-body">
-                    <form action="/cadastro"  method="post" >
+                    <form action="/cadastro-admin"  method="post" >
                         <div class="form-group">
                             <label  for="exampleInputEmail1">
                                 <strong>Nome:</strong>
