+++ App/Views/genero/cadastroGeneros.phtml
@@ -22,7 +22,7 @@
                                 <label for="descricao" class="form-label"> <strong>Descrição: </strong></label>
                                 <textarea class="form-control border border-dark" id="descricao" rows="3" name="descricao"></textarea>
                             </div>
-                            <button type="submit" class="btn btn-dark">Enviar</button>
+                            <button type="submit" class="btn btn-dark">Cadastrar</button>
                         </form>
                     </div>
                 </div>
