+++ App/Views/genero/generos.phtml
@@ -28,7 +28,7 @@
                     <?php endif ?>
 
                     <div class="form-group col-md-12 d-flex justify-content-end">
-                        <a href="/criar-genero" class="btn btn-dark ml-2">Cadastrar gênero</a>
+                        <a href="criar-genero" class="btn btn-dark ml-2">Cadastrar gênero</a>
                     </div>  
                     <div class="table-responsive">
                     <table class="table">
