+++ App/Views/layouts/layoutDashboard.phtml
@@ -39,11 +39,8 @@
                 <li id="btn-musicas">
                     <a href="/minhas-musicas"><i class="fa fa-fw fa-headphones"></i> Músicas</a>
                 </li>
-                <li id="btn-musicas">
+                <li id="btn-generos">
                     <a href="/generos"><i class="fa fa-fw fa-list"></i> Gêneros</a>
-                </li>
-                <li id="btn-cadastro">
-                    <a href="/cadastro"><i class="fa fa-user-circle"></i> Cadastro</a>
                 </li>
                 <li id="btn-admins">
                     <a href="/admins"><i class="fa fa-users"></i> Admins</a>
